Change test-all suite to enable testing each test in a separate ractor.
This tests for ractor safety issues as well as issues with GC and the
ractor scheduler. It also tests move/copy logic because it uses these
internally for running the tests.

You can enable these tests with RUBY_TESTS_WITH_RACTORS=1 when running
`make test-all TESTS=test/ruby`. These ractor tests are currently only
working for tests under the "test/ruby" directory, so don't run any
others with this ENV var set or you will get errors.

Currently there are GC issues with ractors so in tool/lib/test/unit.rb, I disable the
GC before running any of the tests (for now). If you uncomment this, you will get
errors and you can debug the GC issues.

+++ test/ruby/test_argf.rb
@@ -7,6 +7,7 @@ require 'fileutils'
 
 class TestArgf < Test::Unit::TestCase
   def setup
+    omit "ARGF is not shareable" if non_main_ractor?
     @tmpdir = Dir.mktmpdir
     @tmp_count = 0
     @t1 = make_tempfile("argf-foo", %w"1 2", binmode: true)
@@ -15,6 +16,7 @@ class TestArgf < Test::Unit::TestCase
   end
 
   def teardown
+    return if non_main_ractor?
     FileUtils.rmtree(@tmpdir)
   end
